chore: rename user -> applicant
